refactor: 코드리뷰 반영
폴더 구조를 논의한데로 구성 하였습니다.
styled component 순서를 사용한 순서대로 선언순서를 변경하였습니다.
App.js에서 라우트 개행 삭제
body태그 background-color 설정
