+++ src/GlobalStyle.js
@@ -1,4 +1,5 @@
 import { createGlobalStyle } from "styled-components";
+import theme from "./config/constants/theme";
 
 const GlobalStyle = createGlobalStyle`
   * {
@@ -10,6 +11,7 @@ const GlobalStyle = createGlobalStyle`
   #root, body, html {
     width: 100%;
     height: 100%;
+    background-color: ${theme.skyBlue};
   }
 `;
