+++ src/pages/ChannelDetailPage/index.js
